backend/requirements.txt
asgiref
Django
django-debug-toolbar
django-environ
djangorestframework
psycopg2
pytz
sqlparse
